비디오 컨트롤 테스트 추가
