+++ src/main/java/com/numble/whatz/application/video/controller/VideoController.java
@@ -3,7 +3,7 @@ package com.numble.whatz.application.video.controller;
 import com.numble.whatz.application.video.controller.dto.*;
 import com.numble.whatz.application.video.service.VideoService;
 import com.numble.whatz.application.video.service.VideoStore;
-import com.numble.whatz.core.exception.video.CustomVideoStoreException;
+import com.numble.whatz.application.video.service.VideoStoreImpl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
