+++ src/main/java/com/numble/whatz/core/exception/video/CustomVideoStoreException.java
@@ -15,4 +15,6 @@ public class CustomVideoStoreException extends IOException {
         this.videoStoreExceptionMessage = videoStoreExceptionMessage;
         this.e = e;
     }
+
+
 }
