+++ src/test/java/com/numble/whatz/api/video/VideoDocumentationTest.java
@@ -2,13 +2,18 @@ package com.numble.whatz.api.video;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.numble.whatz.api.video.dto.EmbedDto;
+import com.numble.whatz.application.video.service.VideoStore;
+import com.numble.whatz.core.advice.VideoStoreExceptionMessage;
+import com.numble.whatz.core.exception.video.CustomVideoStoreException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.mockito.InjectMocks;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.runner.WebApplicationContextRunner;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockMultipartFile;
@@ -22,8 +27,14 @@ import org.springframework.web.context.WebApplicationContext;
 import org.springframework.web.filter.CharacterEncodingFilter;
 
 
+import java.io.IOException;
+import java.util.Optional;
+
 import static com.numble.whatz.api.utils.ApiDocumentUtils.getDocumentRequest;
 import static com.numble.whatz.api.utils.ApiDocumentUtils.getDocumentResponse;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.doThrow;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.*;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
@@ -44,10 +55,16 @@ public class VideoDocumentationTest {
     @Autowired
     private ObjectMapper objectMapper;
 
+    @MockBean
+    private VideoStore videoStore;
+
     @Test
     public void videoUploadDirect() throws Exception {
         //given
         MockMultipartFile file = new MockMultipartFile("file", "video.mp4", "video/mp4", "<<video data>>".getBytes());
+        doReturn(
+                "success"
+        ).when(videoStore).storeVideo(any());
 
         //when
         ResultActions result = this.mockMvc.perform(
@@ -75,9 +92,13 @@ public class VideoDocumentationTest {
     }
 
     @Test
-    public void videoUploadDirectException() throws Exception {
+    public void videoUploadDirectMultipartException() throws Exception {
         //given
         MockMultipartFile file = new MockMultipartFile("file", "video.mp4", "video/mp4", "<<video data>>".getBytes());
+        doThrow(
+                new CustomVideoStoreException(
+                        VideoStoreExceptionMessage.MULTIPART_EX, new IOException())
+        ).when(videoStore).storeVideo(any());
 
         //when
         ResultActions result = this.mockMvc.perform(
@@ -92,7 +113,75 @@ public class VideoDocumentationTest {
 
         //then
         result.andExpect(status().isForbidden())
-                .andDo(document("video-upload-direct", // (4)
+                .andDo(document("video-upload-direct-multi", // (4)
+                        getDocumentRequest(),
+                        getDocumentResponse(),
+                        requestParts(partWithName("file").optional().description("영상")),
+                        requestParameters(
+                                parameterWithName("videoThumbnail").description("영상 썸네일"),
+                                parameterWithName("title").description("영상 제목"),
+                                parameterWithName("content").description("영상 내용")
+                        )
+                ));
+    }
+
+    @Test
+    public void videoUploadDirectFFmpegException() throws Exception {
+        //given
+        MockMultipartFile file = new MockMultipartFile("file", "video.mp4", "video/mp4", "<<video data>>".getBytes());
+        doThrow(
+                new CustomVideoStoreException(
+                        VideoStoreExceptionMessage.FFMPEG_EX, new IOException())
+        ).when(videoStore).storeVideo(any());
+
+        //when
+        ResultActions result = this.mockMvc.perform(
+                multipart("/api/video/add/direct")
+                        .file(file)
+                        .param("videoThumbnail", "This is Thumbnail")
+                        .param("title", "This is title")
+                        .param("content", "This is content")
+                        .contentType(MediaType.MULTIPART_FORM_DATA)
+                        .accept(MediaType.APPLICATION_JSON_UTF8)
+        );
+
+        //then
+        result.andExpect(status().isForbidden())
+                .andDo(document("video-upload-direct-ffmpeg", // (4)
+                        getDocumentRequest(),
+                        getDocumentResponse(),
+                        requestParts(partWithName("file").optional().description("영상")),
+                        requestParameters(
+                                parameterWithName("videoThumbnail").description("영상 썸네일"),
+                                parameterWithName("title").description("영상 제목"),
+                                parameterWithName("content").description("영상 내용")
+                        )
+                ));
+    }
+
+    @Test
+    public void videoUploadDirectS3Exception() throws Exception {
+        //given
+        MockMultipartFile file = new MockMultipartFile("file", "video.mp4", "video/mp4", "<<video data>>".getBytes());
+        doThrow(
+                new CustomVideoStoreException(
+                        VideoStoreExceptionMessage.S3_UPLOAD_EX, new IOException())
+        ).when(videoStore).storeVideo(any());
+
+        //when
+        ResultActions result = this.mockMvc.perform(
+                multipart("/api/video/add/direct")
+                        .file(file)
+                        .param("videoThumbnail", "This is Thumbnail")
+                        .param("title", "This is title")
+                        .param("content", "This is content")
+                        .contentType(MediaType.MULTIPART_FORM_DATA)
+                        .accept(MediaType.APPLICATION_JSON_UTF8)
+        );
+
+        //then
+        result.andExpect(status().isForbidden())
+                .andDo(document("video-upload-direct-s3", // (4)
                         getDocumentRequest(),
                         getDocumentResponse(),
                         requestParts(partWithName("file").optional().description("영상")),
@@ -114,7 +203,7 @@ public class VideoDocumentationTest {
                 post("/api/video/add/embed")
                         .content(objectMapper.writeValueAsString(embedDto))
                         .contentType(MediaType.APPLICATION_JSON)
-                        .accept(MediaType.APPLICATION_JSON)
+                        .accept(MediaType.APPLICATION_JSON_UTF8)
         );
 
         //then
